frontend/requirements.txt
Flask==2.3.3
python-keycloak==3.7.0
PyJWT==2.8.0
requests==2.31.0